Fixed tape log output for gcc.
(cherry picked from commit a7464e65c3ceb0ec1f0ed53265c1b8f1649cf3ad)

+++ tapescript/cpp/cl/tape/impl/ad/tape_serializer_call.hpp
@@ -225,7 +225,7 @@ namespace CppAD
     inline void serialize__
         (std::true_type
         , Stream                             &s_out
-        , CppAD::vector<unsigned long long>  &user_iy
+        , CppAD::vector<size_t>              &user_iy
         , CppAD::player<Base>*                play
         , Base                               *taylor
         , enum CppAD::OpCode                  op
